tests/wpt/tests/trusted-types/trusted-types-svg-script.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="./support/csp-violations.js"></script>
  <meta http-equiv="Content-Security-Policy"
        content="require-trusted-types-for 'script'">
  <meta http-equiv="Content-Security-Policy" content="object-src 'none'">
</head>
<body>
  <div id="log"></div>
  <svg id="svg"><script id="script">"some script text";</script></svg>
  <script>
    const policy = trustedTypes.createPolicy("policy", {
        createScript: x => x, createHTML: x => x, createScriptURL: x => x });

    promise_test(async t => {
      const input = "'modified via innerHTML';"
      let violation = await trusted_type_violation_for(TypeError, _ =>
        document.getElementById("script").innerHTML = input
      );
      assert_true(violation.originalPolicy.includes("require-trusted-types-for 'script'"));
      assert_equals(violation.sample, `SVGScriptElement innerHTML|${clipSampleIfNeeded(input)}`);
    }, "Assign String to SVGScriptElement.innerHTML.");

    promise_test(async t => {
      const input = "'modified via innerHTML';"
      let violation = await trusted_type_violation_for(TypeError, _ =>
        document.getElementById("script").innerHTML = policy.createHTML(input)
      );
      assert_true(violation.originalPolicy.includes("require-trusted-types-for 'script'"));
      assert_equals(violation.sample, `SVGScriptElement innerHTML|${clipSampleIfNeeded(input)}`);
    }, "Assign TrustedHTML to SVGScriptElement.innerHTML.");

    promise_test(async t => {
      const input = "'modified via innerHTML';"
      const elem = document.createElementNS(
          "http://www.w3.org/2000/svg", "script");
      let violation = await trusted_type_violation_for(TypeError, _ =>
        elem.innerHTML = policy.createHTML(input)
      );
      assert_true(violation.originalPolicy.includes("require-trusted-types-for 'script'"));
      assert_equals(violation.sample, `SVGScriptElement innerHTML|${clipSampleIfNeeded(input)}`);
      document.getElementById("svg").appendChild(elem);
    }, "Assign TrustedHTML to SVGScriptElement.innerHTML and execute it.");

    promise_test(async t => {
      const input = "modified via DOM";
      const elem = document.createElementNS(
          "http://www.w3.org/2000/svg", "script");
      let violation = await trusted_type_violation_for(TypeError, _ =>
        elem.insertBefore(document.createTextNode(input), null)
      );
      assert_true(violation.originalPolicy.includes("require-trusted-types-for 'script'"));
      assert_equals(violation.sample, `SVGScriptElement text|${clipSampleIfNeeded(input)}`);
      document.getElementById("svg").appendChild(elem);
    }, "Modify SVGScriptElement via DOM manipulation.");

    promise_test(t => {
      trustedTypes.createPolicy("default", {
        createScript: (input, _, sink) => {
          assert_equals(sink, 'SVGScriptElement text');
          return input;
        },
        createHTML: (input, _, sink) => {
          assert_equals(sink, 'Element innerHTML');
          return input;
        },
      });
      document.getElementById("script").innerHTML = "'modified via innerHTML';";
      return Promise.resolve();
    }, "Assign String to SVGScriptElement.innerHTML w/ default policy.");
  </script>
</body>
